:truck: Moved OAuth clients into a submodule
