+++ apricot/oauth_clients/enums.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class OAuthBackend(str, Enum):
+    """Available OAuth backends."""
+
+    MICROSOFT_ENTRA = "MicrosoftEntra"
