+++ apricot/oauth_clients/types.py
@@ -1,0 +1,4 @@
+from typing import Any
+
+LDAPEntry = dict[str, list[Any]]
+LDAPEntryList = list[dict[str, list[Any]]]
